refactor: gather code for backend type in one file
