+++ pkg/controller/find/backend.go
@@ -78,13 +78,7 @@ func handleTerraformBlock(block *hclsyntax.Block, bucket *Bucket) (bool, error) 
 			return false, nil
 		}
 		backendType := backend.Labels[0]
-		if backendType != "s3" {
-			return false, nil
-		}
-		handlers := map[string]handleBackend{
-			"s3":  handleS3Backend,
-			"gcs": handleGCSBackend,
-		}
+		handlers := getHandlers()
 		handler, ok := handlers[backendType]
 		if !ok {
 			return false, nil
@@ -98,49 +92,3 @@ func handleTerraformBlock(block *hclsyntax.Block, bucket *Bucket) (bool, error) 
 }
 
 type handleBackend func(backend *hclsyntax.Block, bucket *Bucket) error
-
-func handleS3Backend(backend *hclsyntax.Block, bucket *Bucket) error {
-	bucket.Type = "s3"
-	if key, ok := backend.Body.Attributes["key"]; ok {
-		val, diag := key.Expr.Value(nil)
-		if diag.HasErrors() {
-			return diag
-		}
-		bucket.Key = val.AsString()
-	}
-	if b, ok := backend.Body.Attributes["bucket"]; ok {
-		val, diag := b.Expr.Value(nil)
-		if diag.HasErrors() {
-			return diag
-		}
-		bucket.Bucket = val.AsString()
-	}
-	return nil
-}
-
-func handleGCSBackend(backend *hclsyntax.Block, bucket *Bucket) error {
-	/*
-		terraform {
-		  backend "gcs" {
-		    bucket  = "tf-state-prod"
-		    prefix  = "terraform/state"
-		  }
-		}
-	*/
-	bucket.Type = "gcs"
-	if prefix, ok := backend.Body.Attributes["prefix"]; ok {
-		val, diag := prefix.Expr.Value(nil)
-		if diag.HasErrors() {
-			return diag
-		}
-		bucket.Prefix = val.AsString()
-	}
-	if b, ok := backend.Body.Attributes["bucket"]; ok {
-		val, diag := b.Expr.Value(nil)
-		if diag.HasErrors() {
-			return diag
-		}
-		bucket.Bucket = val.AsString()
-	}
-	return nil
-}
