+++ pkg/controller/find/find.go
@@ -78,12 +78,7 @@ func Find(_ context.Context, logE *logrus.Entry, afs afero.Fs, param *Param) err
 		logE.Info("no backend configuration")
 		return nil
 	}
-	logE.WithFields(logrus.Fields{
-		"bucket": bucket.Bucket,
-		"key":    bucket.Key,
-		"prefix": bucket.Prefix,
-		"type":   bucket.Type,
-	}).Debug("backend configuration")
+	logE.WithFields(bucket.LogE()).Debug("backend configuration")
 
 	// find HCLs in base directories and list directories where changed outputs are used
 	tfFiles, err := findTFFiles(afs, param.Root)
@@ -210,17 +205,6 @@ type Change struct {
 type ChangedFile struct {
 	Path    string   `json:"path"`
 	Outputs []string `json:"outputs"`
-}
-
-type Bucket struct {
-	Type   string `json:"type"`
-	Bucket string `json:"bucket"`
-	Key    string `json:"key"`
-	Prefix string `json:"prefix"`
-}
-
-func (b *Bucket) Compare(bucket *Bucket) bool {
-	return b.Type == bucket.Type && b.Bucket == bucket.Bucket && b.Key == bucket.Key && b.Prefix == bucket.Prefix
 }
 
 type RemoteState struct {
